Explicitly mark PIC register as a save/restore reg.
Bug #4599 (see also #4591).  Even though our PIC register, r24,
is a callee-saved register, its meaning is changed by setting
the call_used_reg[r24] value to 1.  And call_used_reg[r24] is 1
because fixed_regs[r24] is set to 1.  To get around this, we
check for the PIC register explicitly in qpds6_save_register_p.
This is also what the other arch. do.