Bugfix 4609: Fix ABI dependency in gcc source code. The gcc code makes assumptions on how the host ABI behaves. Specifically, it assumes that given a function foo: 1. foo can be declared as a varargs function at the call site 2. However, foo can then be defined as a non-varargs function in another module
For the x86 ABI this works because all parameters are passed on the stack.
For the QDSP6 ABI this fails because within foo's body, it expects the 2nd
parameter to be in r1. However, the call site pushes the 2nd parameter to
stack as per the ABI.

I've changed gcc's source code; specifically, I removed the macro GEN_FCN
and defined GEN_FCNn where n = {1,...,7} and represents the number of
function parameters. I replaced the uses of GEN_FCN with the corresponding
GEN_FCNn macro.

We should change the linux QDSP6 ABI so that the first 6 varargs parameters
are copied to registers in addition to being pushed on the stack. This
allows for an extra level of abstraction between the call site and the
callee.