PIC code for labels in case jump table is incorrect.
Bug 4604.  The code generation for the labels in the jump table
was not generated correctly when using -fPIC.  The problem is that
either all the labels in the table need to be appended with @GOTOFF,
or the calculation for the relative labels in the table need to
use the base of the jump table instead of the pic register.

We opt to use second fix.  Using the base of the jump table to
correct the code labels in the jump table.  This is what many
existing targets do.  As a side effect, we need to embed the jump
table in the text section until bug 4613 is addressed.