Bug 4311. Use better implementation of memcpy_likely_aligned_...
Replaced hexagon_memcpy_likely_aligned_min32bytes_mult8bytes
for Linux.  The patch was supplied as part of bug 4311.  Initally,
the routine was copied here from the dinkumware version since
it is needed for code generation with gcc 4.4.  Some minor changes
were made to get it to work.  A patch was provided that improves
upon the initial version.