Fix bug in generating pic prolog code with hardware loops.
Bug 4837.  The existing code for generating the PIC prolog
code was not working correctly.  The call to require_pic_register
in the hardware loop patterns is too late in the cod gen process
to cause the prolog to be generated.  We need to add the
call to require_pic_register in the define_expand pattern for
hardware loops.  Also, we need to ensure that r24, the pic
register, is live at the loop start.  This will cause the PIC
prolog code to be generate for both loop instruction forms,
unfortunately.