Fix eh_frame return rtx (bug 5066)
C++ exceptions require manipulation of the stack frame to
be able to jump to the exception handler.  The generated
rtx for the exception frame handling is correct, but is
optimized away during compilation. Setting the volatile
flag prevents this from happening, and restores function.

ATTENTION: This patch has NOT (yet) been tested in hexframe,
and it *might* affect operation of the phone code.