Unwind stack past a frame containing a linux signal tramp.
eh_frame unwinding in libgcc will fail if a stack-frame contains
a linux signal trampoline. This patch adds the needed support;
it checks for the presence of the tramploine, and if found,
it sets up the dwarf2 "columns" to point at the correct locations
of the assorted registers that were saved in the frame.