Need seperate patterns for pic hardware loop insns (Bug 5105).
Generating pic code for a hardware loop uses the r24 (pic/got
register) when the compiler generates code to setup the sa0/lc0
register explicitly.  However, the pattern doesn't explicitly
indicate that the r24 register is used.  This causes a problem
because the pic setup code can be moved to after the loop0
sequence (see also Bug 4837).  We had tried to fix this by adding
a use of r24 prior to the loop0 insn.

This test case shows that the prior fix doesn't work because
the schedule phase may move the use of r24 after the loop0 insn.
To fix this, we need to add the use of r24 to the loop0 insn
explicitly.  We do this by creating seperate patterns for pic
loop0 instructions, which allows us to specify that r24 is used
only when pic code is generated.  The non-pic case will not
be affected.