Save PIC reg, if saves_all_registers flag is set.
See bug 5173 for details.
The exception unwinding code makes use of __builtin_unwind_init(),
which should save all non-volatile regs.  However, on hexagon,
we were failing to save r24, as, for this case, it isn't live,
and because, then pic_flag is set, its marked as a fixed reg.
Solution is to save it if the saves_all_registers flag is set.