Bug 5314. Don't generate hw loops if no. of iterations may overflow.
We should not generate a hardware loop if the number of runtime
computed loop iterations may be UINT_MAX.  On Hexagon, we increment
the number of iterations by 1, which overflows the counter and the
number of iterations becomes 0 (and that really means 1).  This
case may occur only when the loop exit condition is at the start
of the loop.  In cases where the exit condition is at the end,
this condition will not occur.

We check the condition in the target independent code for
hardware loops since there is no convenient target dependent hook
for this.

Same as bug 5262.

Change-Id: Ie62dcce0c91f9374f8c30a7a2136a5eada76255e